requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.7.0
python-multipart==0.0.9

# LLM providers — at least one required
groq==0.9.0
openai==1.35.0
anthropic==0.28.0